requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.7.4
decorator==5.1.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
flake8==7.1.1
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
mccabe==0.7.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytils==0.4.1
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
